mod: grouping api

+++ query_service/query_api/crm/entity/dto/grouping/dto.py
@@ -257,9 +257,9 @@ cml_consumption_months = ns_5.model('CumulativeConsumptionMonthsConditionModel',
 })
 
 cml_consumption_amount = ns_5.model('CumulativeConsumptionAmountConditionModel', {
-    'lt': fields.Integer(description="小于", example=2, min=0),
-    'gt': fields.Integer(description="大于", example=2, min=0),
-    'eq': fields.Integer(description="等于", example=2, min=0),
+    'lt': fields.Float(description="小于", example=2, min=0),
+    'gt': fields.Float(description="大于", example=2, min=0),
+    'eq': fields.Float(description="等于", example=2, min=0),
     'bt': fields.List(fields.Integer(description="介于", min=0), example=[0, 5]),
 })
